Renaming 'vecmath' to 'colorvec'
The new name is not only more descriptive (after all, one of the points
of it is to have both 'rgb' and 'xyz' as channel/component names), but
it also will help to avoid confusion if anyone else has a 'vecmath'
library.

This commit is 'part 1', and only renames the files. The next commit
will actually change the contents of those files to reflect the change.