Reduce C++ version dependency from C++17 to C++11
There was some bizarre bug in the past relating to the tests that forced
me to upgrade to C++17. Still don't understand why it behaved the way it
did.

But now it works with just C++11 again, and maybe earlier versions? I'm
not going to test those quite yet, C++11 is still relatively new, but
also fairly well supported. I'll just stick to that for now.