Added a LOT of stuff (including some Qt widgets)!
Things are still a bit rough witw the Qt widgets, but it's more of a
proof-of-concept than anything else right now.

Beyond that, now we've split up color_space a bit and also added a class
for specifying transfer characteristics (TRCs). In the future there will
be log-based TRCs as well, and I also plan on including the standard for
HDR, 'Hybrid Log-Gamma' (HLG).